Move district drawing out of projectData
