+++ src/client/actions/projectData.ts
@@ -29,22 +29,3 @@ export const staticDemographicsFetchSuccess = createAction("Static demographics 
 export const staticDemographicsFetchFailure = createAction("Static demographics fetch failure")<
   string
 >();
-
-export const setSelectedDistrictId = createAction("Set selected district id")<number>();
-
-export const addSelectedGeounitIds = createAction("Add selected geounit ids")<
-  ReadonlySet<number>
->();
-export const removeSelectedGeounitIds = createAction("Remove selected geounit ids")<
-  ReadonlySet<number>
->();
-export const clearSelectedGeounitIds = createAction("Clear selected geounit ids")();
-
-export const saveDistrictsDefinition = createAction("Save districts definition")();
-
-export const patchDistrictsDefinitionSuccess = createAction("Patch districts definition success")<
-  IProject
->();
-export const patchDistrictsDefinitionFailure = createAction("Patch districts definition failure")<
-  string
->();
