+++ src/client/reducers/projectData.ts
@@ -4,10 +4,6 @@ import { getType } from "typesafe-actions";
 
 import { Action } from "../actions";
 import {
-  addSelectedGeounitIds,
-  clearSelectedGeounitIds,
-  patchDistrictsDefinitionSuccess,
-  patchDistrictsDefinitionFailure,
   projectDataFetch,
   projectFetch,
   projectFetchFailure,
@@ -15,9 +11,6 @@ import {
   projectFetchGeoJsonFailure,
   projectFetchGeoJsonSuccess,
   projectFetchSuccess,
-  removeSelectedGeounitIds,
-  saveDistrictsDefinition,
-  setSelectedDistrictId,
   staticDemographicsFetchFailure,
   staticDemographicsFetchSuccess,
   staticGeoLevelsFetchFailure,
@@ -27,7 +20,7 @@ import {
 } from "../actions/projectData";
 
 import { DistrictProperties, IProject, IStaticMetadata } from "../../shared/entities";
-import { fetchProject, fetchProjectGeoJson, patchDistrictsDefinition } from "../api";
+import { fetchProject, fetchProjectGeoJson } from "../api";
 import { Resource } from "../resource";
 import { fetchStaticFiles, fetchStaticMetadata } from "../s3";
 
@@ -37,8 +30,6 @@ export interface ProjectDataState {
   readonly staticGeoLevels: Resource<ReadonlyArray<Uint8Array | Uint16Array | Uint32Array>>;
   readonly staticDemographics: Resource<ReadonlyArray<Uint8Array | Uint16Array | Uint32Array>>;
   readonly geojson: Resource<FeatureCollection<MultiPolygon, DistrictProperties>>;
-  readonly selectedDistrictId: number;
-  readonly selectedGeounitIds: ReadonlySet<number>;
 }
 
 export const initialState = {
@@ -46,13 +37,9 @@ export const initialState = {
   staticMetadata: { isPending: false },
   staticGeoLevels: { isPending: false },
   staticDemographics: { isPending: false },
-  geojson: { isPending: false },
-  selectedDistrictId: 1,
-  selectedGeounitIds: new Set([])
+  geojson: { isPending: false }
 };
 
-// TODO (#187): this reducer is getting large. Some of the actions should be split out
-// into a separate module that deals more with district selection.
 const projectDataReducer: LoopReducer<ProjectDataState, Action> = (
   state: ProjectDataState = initialState,
   action: Action
@@ -184,67 +171,6 @@ const projectDataReducer: LoopReducer<ProjectDataState, Action> = (
           errorMessage: action.payload
         }
       };
-    case getType(setSelectedDistrictId):
-      return {
-        ...state,
-        selectedDistrictId: action.payload
-      };
-    case getType(addSelectedGeounitIds):
-      return {
-        ...state,
-        selectedGeounitIds: new Set([...state.selectedGeounitIds, ...action.payload])
-      };
-    case getType(removeSelectedGeounitIds): {
-      const mutableSelected = new Set([...state.selectedGeounitIds]);
-      [...action.payload].forEach(function(v) {
-        mutableSelected.delete(v);
-      });
-      return {
-        ...state,
-        selectedGeounitIds: mutableSelected
-      };
-    }
-    case getType(clearSelectedGeounitIds):
-      return {
-        ...state,
-        selectedGeounitIds: new Set([])
-      };
-    case getType(saveDistrictsDefinition):
-      return "resource" in state.project
-        ? loop(
-            state,
-            Cmd.run(patchDistrictsDefinition, {
-              successActionCreator: patchDistrictsDefinitionSuccess,
-              failActionCreator: patchDistrictsDefinitionFailure,
-              args: [
-                state.project.resource.id,
-                // TODO (#113): we are only dealing with the top-most geolevel at the moment, so this
-                // will need to be modified when we support all geolevels.
-                [...state.selectedGeounitIds].reduce((newDistrictsDefinition, geounitId) => {
-                  // @ts-ignore
-                  // eslint-disable-next-line
-                  newDistrictsDefinition[geounitId] = state.selectedDistrictId;
-                  return newDistrictsDefinition;
-                }, state.project.resource.districtsDefinition)
-              ] as Parameters<typeof patchDistrictsDefinition>
-            })
-          )
-        : (state as never);
-    case getType(patchDistrictsDefinitionSuccess):
-      return loop(
-        {
-          ...state,
-          project: { resource: action.payload },
-          selectedGeounitIds: new Set([])
-        },
-        Cmd.action(projectFetchGeoJson(action.payload.id))
-      );
-    case getType(patchDistrictsDefinitionFailure):
-      // TODO (#188): implement a status area to display errors for this and other things
-      // eslint-disable-next-line
-      console.log("Error patching districts definition: ", action.payload);
-      return state;
-
     default:
       return state as never;
   }
